requirements.txt
Django==2.1.2
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0
Pillow==5.3.0
django-extra-fields==1.0.0
django-multiselectfield==0.1.8
requests==2.19.1
feedparser==5.2.1
python-dateutil==2.7.3
django-rest-swagger==2.2.0
gspread==3.0.1
oauth2client==4.1.3
django-notifications-hq==1.5.0
pywebpush==1.7.0
pyfcm==1.4.5
beautifulsoup4==4.6.3
freezegun==0.3.10
django-markdownify==0.8.0
